07/requirements.txt
openai>=1.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
